+++ lib/bamboo/adapters/mandrill_adapter.ex
@@ -19,44 +19,13 @@ defmodule Bamboo.MandrillAdapter do
       end
   """
 
+  @service_name "Mandrill"
   @default_base_uri "https://mandrillapp.com"
   @send_message_path "api/1.0/messages/send.json"
   @send_message_template_path "api/1.0/messages/send-template.json"
   @behaviour Bamboo.Adapter
 
-  defmodule ApiError do
-    defexception [:message]
-
-    def exception(%{message: message}) do
-      %ApiError{message: message}
-    end
-
-    def exception(%{params: params, response: response}) do
-      filtered_params = params |> Poison.decode! |> Map.put("key", "[FILTERED]")
-
-      message = """
-      There was a problem sending the email through the Mandrill API.
-
-      Here is the response:
-
-      #{inspect response, limit: :infinity}
-
-      Here are the params we sent:
-
-      #{inspect filtered_params, limit: :infinity}
-
-      If you are deploying to Heroku and using ENV variables to handle your API key,
-      you will need to explicitly export the variables so they are available at compile time.
-      Add the following configuration to your elixir_buildpack.config:
-
-      config_vars_to_export=(
-        DATABASE_URL
-        MANDRILL_API_KEY
-      )
-      """
-      %ApiError{message: message}
-    end
-  end
+  import Bamboo.ApiError
 
   def deliver(email, config) do
     api_key = get_key(config)
@@ -65,11 +34,12 @@ defmodule Bamboo.MandrillAdapter do
 
     case :hackney.post(uri, headers(), params, [:with_body]) do
       {:ok, status, _headers, response} when status > 299 ->
-        raise(ApiError, %{params: params, response: response})
+        filtered_params = params |> Poison.decode! |> Map.put("key", "[FILTERED]")
+        raise_api_error(@service_name, response, filtered_params)
       {:ok, status, headers, response} ->
         %{status_code: status, headers: headers, body: response}
       {:error, reason} ->
-        raise(ApiError, %{message: inspect(reason)})
+        raise_api_error(inspect(reason))
     end
   end
 
@@ -81,6 +51,9 @@ defmodule Bamboo.MandrillAdapter do
       config
     end
   end
+
+  @doc false
+  def supports_attachments?, do: true
 
   defp get_key(config) do
     case Map.get(config, :api_key) do
@@ -119,7 +92,8 @@ defmodule Bamboo.MandrillAdapter do
       subject: email.subject,
       text: email.text_body,
       html: email.html_body,
-      headers: email.headers
+      headers: email.headers,
+      attachments: attachments(email)
     }
     |> add_message_params(email)
   end
@@ -130,6 +104,18 @@ defmodule Bamboo.MandrillAdapter do
     end)
   end
   defp add_message_params(mandrill_message, _), do: mandrill_message
+
+  defp attachments(%{attachments: attachments}) do
+    attachments
+    |> Enum.reverse
+    |> Enum.map(fn(att) ->
+      %{
+        name: att.filename,
+        type: att.content_type,
+        content: Base.encode64(File.read!(att.path))
+      }
+    end)
+  end
 
   defp recipients(email) do
     []