+++ test/lib/bamboo/email_test.exs
@@ -76,4 +76,11 @@ defmodule Bamboo.EmailTest do
 
     assert email.private["foo"] == "bar"
   end
+
+  test "put_attachment/3 atts an attachment to the attachments list" do
+    path = Path.join(__DIR__, "../../support/attachment.docx")
+    email = new_email() |> put_attachment(path)
+
+    assert [%Bamboo.Attachment{filename: "attachment.docx"}] = email.attachments
+  end
 end